switched from Controllers to Controller class
