+++ lib/content_editor.rb
@@ -1,7 +1,5 @@
-# def replace_content(replacement: {}, path: "")
 def replace_content(path: "")
   File.open(path, "w") do |f|
-    # f.puts render(replacement: replacement)
     f.puts render(path: path)
   end
 end
